Adjust "Goto Command" box to support specifying a .-delimited subcommand
We're about to have the replay profiler slices that link to specific
subcommands. A reasonable step along the way is to be able to specify
subcommands using this dialog.

Converts the spinner widget to a textbox.

Navigating to a subcommand doesn't actually work yet -- we need to build
out a bit more backend plumbing for that, but this is a first step.

Bug: b/143735500