Use Perfetto Client API in tracing.
Previously we used Perfetto CLI to trace, however, when tracing is started,
some ftrace are not set up yet, which results in a gap of empty data at the
beginning of the trace. With Perfetto Client API, we now can do deferred start
by setting deferred_start field to true, and always make sure we only start the
tracing session when things are ready.

Bug: b/147490759