Keep only one per-thread persistent CallObserver
This change removes the code related to stacking of call
observers (e.g. the logic involving mParent): that stacking was a
reminiscence of a graphics API (GVR I think) needs, where commands
could call other commands in a nested way.

In addition, rather than creating a new CallObserver each time we
enter a command, this change also use persistent, per-thread call
observers. They are stored in the thread-local `tl_callObserver`,
which used to be named `gContext`. This avoids to create and destroy a
CallObserver for each observed command.

This is based on a cherry-pick of
https://github.com/AWoloszyn/gapid/commit/6f66f91fe0b2c641d62d847042c82e555aea7ae8

Bug: b/151400181
Test: manual