Create only one staging buffer per device
When retrieving the Vulkan state at the start of a mid-execution
capture, we create a new staging buffer for each of the buffers in the
state.

This change creates only one staging buffer per device, and reuses it
to retrieve all buffers of that device.

To enable using the make_unique definition found in
memory_tracker_layer_impl.cpp everywhere, we move it into
core/cc/make_unique.h

Bug: none
Test: capture a few Vulkan apps

Apply review comments

Move make_unique