Events without mutate
We only needed to mutate for:
- command filter on events, which we don't use
- possible unspecified extension uses, of which we have none
- the old GLES transform feedback / drawcall flags

Unpicking this allows us to remove the mutate (and dependency on sync
data) making events significantly cheaper.

This reduces the cost of building the event list on the Crytek trace
from 540ms to 1.2ms on a Z840. Total cost of building the command tree
on the same machine reduces from 2.3s to 1.8s. No significant impact to
wall clock time to idle.