Apply review comments
Make Devices.loadReplayDevice return an empty list (and not null) when
the replay device loading succeeded but did return an empty list. This
enables to monitor when the replay devices have been loaded (before,
replayDevices was null if either devices were not loaded yet, or there
was no replay devices available).

Also, DeviceDialog is not a composite anymore.