Always use the device memory pool for linear images.
We typically store image data in image owned pools in the tightly packed
format. This makes it easy to "get" the data for state reconstruction,
the UI, image priming, etc. This is fine for optimally tiled images,
since there the data layout in the device memory is not defined
(implementation owned) and never visible to the application. However,
for linear tiled images the device memory layout is defined and it can be
made visible to the application. In some situations (e.g. pre-initialized,
non-TRANSFER_DST images) we need the linear layed-out data. So, to simplify
the logic of when the data is linearly-layed out, always use the linear
layout (i.e. the device memory pool) for linear tiled images.