Fix image copy when the image size is not a multiple of the block size.
When copying to/from compressed images, the regions to copy need to be
block aligned both in position and size. However, there is one exception
called out in the spec: if the image size is not a multiple of the block
size, then the last block on each row as well as the entire last row can
be "unaligned" and incomplete. To simplify the logic, since we expect
validation-passing apps, we simply always round up the size.