Make ipRenderKitBuilder.Free() more deterministic
Replace some maps by a pair of map + slice in order to be able to free
them in order. Prefer this to sorting keys because here the key type
is a structure with many fields, and defining the "less than" function
would have add a lot of code complexity. The approach here is just to
store them in some order, and reuse this order to free them.