Frame lifecycle UI Updates
We have decided to revamp the frame lifecycle UI to be friendly towards
users who don't have the knowledge of SurfaceFlinger's events. This new
change brings in an ownership view where the old events are taken and
are shown as slices of DISPLAY (Present to Present), APP (Dequeue to
Queue), Wait for GPU (Queue to Acquire), Composition (Latch to Present).
The old buffer events will still be made available so as to not block
experienced users.

This change also removes the old summary view as it is no longer
relevant and there isn't a design yet to show some kind of a summary.

Bug: 155242423