Harden GAPIR error reporting
Avoid early returns upon errors in the ReplayHandler. These early
returns would lead to a hang. The ReplayHandler executes while the
main replayer logic waits on server->wait(), which effectively waits
on the gRPC connection to terminate. When the ReplayHandler handles an
error via an early return, this leads to the main thread waiting on
the connection to terminate, which will never happen: GAPIS is waiting
for a replay response and will not close the connection, the replayer
is waiting for the connection to terminate, no one will make progress.

Also, set a deadline in gRPC Shutdown(), to bound the termination of
in-flight RPCs.

Bug: b/158316277
Test: Lots of manual tests by hard-coding fake errors in e.g. state
priming.