Simplify gapir socket naming
The socket used for GAPIS-GAPIR connection has an ABI-specific name:
this is not needed, because it is backed by a filesystem socket that
is created in the APK directory on device, and there is a distinct APK
for each ABI, so the name of the directory already makes it unique.

This change reduces the number of const strings that have to be kept
in sync between Go and C++ (here socket names), and properly documents
that these strings should be kept in sync.