Cleanup gapir/client package
This is a significant cleanup of the gapir/client package that should
not change the package behavior. The code essentially behaves in the
same way, it just has less indirections, less functions, less types,
and less interfaces. Function, type and package documentation has been
corrected and updated, some unused pieces of code have been removed,
and some functions, types and variables have been renamed to better
reflect what they represent.

This change is better reviewed as a clean reading of the client.go
file, which now contains the logic that was distributed across
client.go, background_connection.co and connection.go. Again, this
change does not change the code logic: there are not-so-good
interactions left over (like the fact that the Replay RPC stream is
created inside handleReplayCommunication() and creation status is
reported back via a channel, when it could be simply created in
Client.Connect() as part of a replayer initialization), for the sake
of preserving the behavior. This is a cleanup to have a cleaner state
from which we can achieve actual behavior refactoring.

Refactoring notes:

- the information related to a single GAPIR instance was stored across
the clientInfo, connection and backgroundConnection types, with some
duplicated fields. These three types are merged into the single
replayerInfo type, which can be seen as the spiritual descendant of
clientInfo that has absorbed connection and backgroundConnection.

- clientInfo (now replayerInfo) stored both a *device.ABI and
device.Architecture field, the architecture field has been removed
because it is trivially retrievable from the ABI field.

- an auth.Token was stored in deviceConnectionInfo and replicated in
connection, it is now stored only once in deviceConnectionInfo.

- the OS field of backgroundConnection is trivially retrievable from
the bind.Device field, hence the OS is not stored as a separate field.

- type deviceArch was declared only to be aliased to the exported
ConnectionKey type, we now just declare a ReplayerKey type that
effectively replaces ConnectionKey.

- the 'conn' variable name was abused to refer to either a connection
or a ConnectionKey, now all ReplayerKey variables are properly named
'key'.

- the Connection and ReplayResponseHandler interfaces defined in
api.go were not used outside of the gapir/client package. The actual
gapir-related functions used outside of gapir/client are the ones
exported by the Client type (namely New, BeginReplay, Connect,
PrewarmReplay, and SetReplayExecutor), and the ReplayExecutor
interface. Having the Connection and ReplayResponseHandler interface
being declared in gapir/api.go was somehow misleading in indicating
that these interfaces were relevant outside of the gapir/client
package. These two interfaces became obsolete and are now removed.

- most exported functions of Client receive a ReplayerKey: the lookup
of this key, and the subsequent check of the replayer found, was
duplicated in various places, it is now factored in
Client.getActiveReplayer().

- the former client.ping() function returned a time.Duration which was
never used, this time.Duration return value is removed.

- some small functions used only once have been inlined.

- some functions (e.g. BeginReplay and PrewarmReplay) were defined
three times for types Client, connection and backgroundConnection,
with the backgroundConnection being a wrapper around the connection
one, and Client being a wrapper around the backgroundConnection
one. This lead to intricate call hierarchies and redirections that are
hard to reason about. With the unification of types under
replayerInfo, most of these duplicates are gone.