Remove Graphics driver override.
Previously we added an override to allow fast development and iteration
with partners, now that the integration is done, remove this hack.
Rationale: Graphics driver has special permission than other components,
for example application apk. Having a way to override this exposes some
security concerns and we will trap ourselves by having to do hacks in
order to make the namespace work properly.

Minor: Do not return the prerelease driver if it's preinstalled, because
it is empty.

Bug: b/148970509