Add support to use GPU profiling libraries in system image.
Previously GPU profiling libraries are shipped as part of developer driver. In
Android 11, now vendors have the choice to ship them as part of BSP. This patch
below supports:

1) Load libgpudataproducer.so from BSP;
2) Add another field to store native GPU profiling libraries;
3) Allow to capture a trace with BSP GPU profiling libraries;
4) Allow to run gapit validation command against the BSP.

Minor: 1) Move frame lifecycle out of GPU profiling capability to Perfetto
capability because it's technically not updatable and bound to the system
image; 2) Change launch_producer to agi_launch_producer so that binary name is
aligned everywhere.

Bug: b/148970509, b/148950543
Test: Verified with  Pixel 4 and Note 10.
Test: bazel run gapit -- validate_gpu_profiling --os android
Test: bazel run gapit -- validate_gpu_profiling --os android --usesystemimage