Move the hover card position to the left of the hover point.
Previously when hover point moves towards to the right boundary of the window,
the hover card will stick on the right edge of the window which blocks the view
of the current hover point and hence the selection. This patch modifies the
position of the hover card and moves the position of the hover card to the left
side of the hover point.

Bug: b/158506210