Avoid "Viewing full screen" notifications
Swarming devices typiclaly have fresh Android installs with all the
first-time-use warnings. Some swarming tests seemed to fail due to the
app loosing focus because of a "Viewing full screen" notification
taking over and waiting for user input. This change makes sure to
disable this notification when preparing the device.

Bug: b/161212931
Test: using test/swarming/manual-run.sh