Better horizontal scrolling in the system profiler.
When the scrolling is done via the touch pad or mouse wheel, handle
it ourselves, rather than going via the scrollbar. The scrollbar
size is limited, and thus when zoomed in, the scroll accuracy is
limited and leads to scrolling too fast/far. We can scroll better
by handling it directly if the event comes from the touch pad or
mouse wheel.