Improve replay device selection dialog
Improve replay device selection dialog by diplaying all known devices
in two tables: compatible devices which can be used for replay, and
incompatible devices, alongside the incompatibility reason to explain
for each why it is not fit to replay the trace.

This change introduces a new "Incompatibility" protobuf enum, and
augment the Devices proto message with a list of IncompatibleDevice,
such that when GAPIS is queried for the list of devices available to
replay a given trace, it also returns the list of incompatible
devices, along with the incompatility reason.

On the GAPIS side, the incompatibility reason is set in
gapis/api/vulkan.GetReplayPriority(), and propagated to the answer to
a GetDevicesForReplay() RPC response.

On the GAPIC side, models.Devices.loadReplayDevices() handles the
augmented RPC reply to store both compatible and incompatible
devices. Each set of devices is presented in a table in
views.DeviceDialog. Selecting a compatible replay device triggers a
validation check, and if validation pass then the user can open the
trace with the selected device.