Make sure the correct error is shown if the mesh is empty.
The UI was using the DataUnavailableException, which is assumed to mean
a non-draw call was selected. There is a TODO to make that exception
more semantically meaningful, already. This change just fixes it so that
this error does show up, when a draw call is selected, but the Mesh is
unavailable (e.g. for samples where the vertex data is part of the
shader).