Fixes a memory alignment issue.
The crash here was due to us trying to read/write 256 bytes from a 64
byte buffer. The 256 comes from the nonCoherentAtomSize constant. When
using the scratch resources buffer we align things to 256 bytes defensively
for the conherent memory case. We, however, are also re-using this code
for the linear image priming, where we do not need that alignment. When
the code was made to be re-used, the alignment was turned into a parameter
to the functions, but here it was missed and stayed an inline constant.