Fix issue where traces weren't properly opened.
The device dialog had a race condition. After loading a capture, the UI
is scheduled to be created, while at the same time the replay devices are
scheduled to be loaded. If the replay devices finished loading before the
UI was created, the device dialog never got the "devices loaded" event.
This changes fixes it by having the dialog initialize itself based on the
current state of the models when it is created, instead of relying on the
timing of its creation.