Make sure to report devices that do not support profiling
Android devices that do not support GPU profiling should still be
listed by GAPIS. In particular, this is needed to list them in the UI
when selecting a replay device: they will be listed as invalid
devices, but at least the user can see that AGI is aware of these
devices.

In practice, this change refactors the GPU profiling setup functions,
and makes sure gapidapk.fetchDeviceInfo() does not return an error
when there is not GPU profiling support.

We had a very similar code duplicated in gapidapk.fetchDeviceInfo(),
gapidapk.ensurePerfettoProducerLaunched() and
gapis/perfetto/android/trace.go:Start() to query the support of GPU
profiling and setup the device accordingly. This change refactors this
code in a single new function: adb.Device.PrepareGpuProfiling().

This new function does both profiling support query and device setup,
since we never query without doing the setup if support is
detected. Both HasGpuProfilingSupportInSystemImage() and
GetGpuProfilingLayerPackageName() become obsolete and are inlined in
the new function, since this is the only place where they are used.

Profiling support is not checked in ensurePerfettoProducerLaunched()
since this function is only called from fetchDeviceInfo(), and only if
support has already been checked.

Bug: b/161878283
Test: manual