Fix undefined pool ID sometimes panicking state rebuilder
Initialize gapii spy mNextPoolId to 1, such that observations recorded
when encoding the state at the beginning of a mid-execution capture
have relevant pool IDs. As this field was not initialized, it had an
undefined value which in practice was often either:

 - 0, such that the first observation related to state encoding would
   end up in the application pool ID 0 -- while this is bogus, it
   seemed to be working fine in most cases.

 - -1 (0xffffffff), such that the second observation would have a pool
   ID of 0, leading to a panic in the state rebuilder when trying to
   create this pool ID 0 since this pool ID is already present as the
   application pool (the panic is in gapis/memory.Pools.NewAt())

Also, improve the panic error message to indicate the requested pool
ID, and fix a typo.

Bug: b/157626574
Test: manual, with an app that exposed the panic quite regularly