Fix GAPIS termination deadlock
Make the error channel `c` be asynchronous to prevent a deadlock in
the termination sequence of GAPIS. The deadlock is related to the
Status() RPC, where a write to the synchronous `c` channel is blocking
since the corresponding read cannot be reached before this write has
terminated.

In gapis/server/grpc.go:grpc.Status(), the synchronous channel `c` is
created and used in the `f` lambda, which writes into it. This `f` is
passed to the call to s.handler.Status(), and this call must terminate
with a non-nil error to have `c` being read from.

It turns out the s.handler.Status() call leads to a blocking write to
`c`. Looking at its implementation in
gapis/server/server.go:server.Status(), note the status.Finish()
deferred function registered at the beginning, then the registration
of a new statusListener that embeds `f`. When the context is
cancelled, this Status() function wants to terminate, of the defered
status.Finish() is called, triggering statusListener.OnTaskFinish(),
which itself calls `f`, which tries a stream.Send(). If this
stream.Send() fails, `f` blocks on a write to `c`, effectively
preventing the termination of server.Status(). However, sever.Status()
must return for grpc.Status() to have a chance to read from `c`.

Making `c` asynchronous avoid `f` to block on the write, thus letting
sever.Status() terminate such that grpc.Server() can also terminate.

Bug: b/150773653
Test: manual, make sure AGI always terminates properly, the deadlock
is almost systematic.