Avoid recording invalid command buffer
Some of the command buffers contained in the initial state cannot be
recorded due to some lacking resources. This fix adds a check to avoid
recording a command buffer which has a vkCmdExecuteCommands command
with as an argument a list of secondary command buffers where one of
the secondary command buffer recording has not been completed.

Also, fix a variable name typo: use `cb` to refer to a command
buffer (`qp` probably came from a copy-paste of the loop on QueryPools
just above).

Bug: b/161071991

Test: manual, with a trace that had a vkCmdExecuteCommands with a
secondary command buffer whose recording was not completed.