Pass along the context for function-based resolves.
For functions stored in the DB to compute the result of a resolve, pass
along a context to the function invocation. Since these functions often
rely on resolving another object themselves, they require a context. These
functions would thus end up holding references to out-of-date contexts,
and status updates would get attributed to the wrong location.