Fix possible replay crash in vkBeginCommandBuffer
Add a read observation on inheritanceInfo even for primary command
buffers.

If at capture time the pInheritanceInfo pointer is non-nil yet this
observation is not made (because the command buffer is primary), at
replay time the value pInheritanceInfo will be `0xbadf00d`, which is
the magic value used in GAPIS to mark pointers that should not be
dereferenced. This should be fine as this command buffer is primary,
so pInheritanceInfo should be ignored. Still, it appears that
validation layers try to dereference it even for primary command
buffers, leading to a replay segfault (for an app that requests
validation layers at capture time, such that any replay also has the
validation layers requested). This might be an issue in the validation
layers, but in any case let's make AGI robust to this by adding the
observation of pInheritanceInfo as soon as the pointer is not null.

Bug: b/161126544

Test: manually, using the com.powervr.VulkanPVRScopeExample demo that
systematically lead to the crash described above.