Add scaffolding for the framegraph feature
This adds the scaffolding code to start working on the framegraph
feature: a gapit framegraph verb, a GetFramegraph RPC, and a
framegraph module.

The code here always returns an empty framegraph: this is on purpose,
the actual creation of a meaningful framegraph will be added in
subsequent changes, once this scaffolding is in place.

Bug: b/171694399