Framegraph: record renderpasses and their dependencies
Add the structure to go over subcommands, record renderpasses, and
establish their dependencies based on the dependency graph. For now,
renderpasses have very little information attached to them: this is on
purpose, to keep this commit small.

The main part of this commit is the inter-renderpass dependencies
logic, in updateDependencies().

The change in MutateWithSubcommands() is to allow passing a nil
postCmdCb to it.

Bug: b/172218304
Test: manually, on a few traces.