Framegraph: add compute workloads
We used to collect information only for renderpasses, this change adds
information gathering for compute workloads as well.

The key change is the renderpassInfo type becoming workloadInfo: it
now stores info that are common to a renderpass and a
compute (e.g. image and buffer accesses), and has a pair of pointer
fields to store renderpass-specific or compute-specific info. The
rpInfo field of helpers is accordingly renamed wlInfo, which leads to
a lot of small-and-trivial-renaming diffs.

Two other noticeable changes:

- In processSubCommand, the detection of renderpass start/end used to
  rely on catching vkCmdBeginRenderpass and vkCmdEndRenderpass: this
  is now a bit more complex since we need to also detect the start/end
  of compute workloads, which can be any of `vkCmdDispatch*`.

- We now list image usage bits as separate booleans in the image
  proto, this mirrors what we do for buffers and leads to a more
  homogeneous representation of images and buffers in framegraph data.
  This effectively replaces the FramegraphImageNature message.

Bug: b/173007235
Test: manual