Update SPIRV-Reflect to fix risk of infinite recursion
When tracing some apps, we hit a SPIRV-Reflect issue that triggers an
infinite recursion and leads to a stack overflow:

https://github.com/KhronosGroup/SPIRV-Reflect/issues/77

This has been fixed in the latest version of SPIRV-Reflect.

Note that we are now tracking SPIRV-Reflect from the Khronos
repository. Moreover, this project now ships its own BUILD.bazel file,
some include paths need to be updated to match this new BUILD.bazel
file.

Bug: b/172066817
Test: manual, on an app that always triggered this infinite recursion.