Don't stop pinging the server upon ping failure.
Changes the heartbeat to keep on pinging the server, since this is a
"I'm still here!" kind of heartbeat, as opposed to an "are you still
there" kind. If the server temporarily gets too busy to answer pings
in time, or the client too busy to process results, we shouldn't stop
sending pings, as when things go back to normal, the server would
eventually quit on a longer idle period.