Framegraph: fix reporting of mapped memory
This change sets the memoryMapped flag of framegraph buffers/images
only when the mapped memory is actually overlapping a range of memory
that backs the buffer/image.

Also, when updating the buffer/image lookup tables, make sure to also
scan the sparse memory of each resource.

Also, stop assuming that there is always a non-sparse memory bound,
this removes some possible NPE.

Bug: b/175394793, b/174505556
Test: manual