Allow installed apk to be queryable.
Starting in Android 11 the package visibility is changed per
https://developer.android.com/training/basics/intents/package-visibility, for
applications that target Android 11 or beyond will no longer be able to query
all packages by default. This prevents our apk from being discovered in the
application process and hence the graphics environment is not properly set up
and the application process can not load any Vulkan layer inside the apk. This
patch adds --force-queryable as one of the args when apk is installed to allow
the apk to be queryable by other applications.

Bug: b/175513065
Test: Run with an application targets api 30, no AppFilters error