Remove nested frame concept in the spy
I think the nested frames only made sense in GLES. In any case, they
are obsolete in Vulkan.

mNestedFrameStart is a leftover from the old @frame_start annotation.

In Vulkan, only vkQueuePresentKHR is annotated with @frame_end, so
upon this call the spy triggers Spy::onPreEndOfFrame() and
Spy::onPostEndOfFrame(), which respectively increments and decrements
the mNestedFrameEnd. However, since vkQueuePresentKHR is not marked as
@threadsafe, there is no chance for concurrent calls of
vkQueuePresentKHR, thus when processing a vkQueuePresentKHR call
mNestedFrameEnd is always incremented to one, then decremented to
zero, such that we can effectively get rid of this counter and its
associated logic.

Bug: b/156105650 (preparatory cleanup)